+++ src/SYSTEM/neighbors.f90
@@ -3,7 +3,7 @@ subroutine neighbors ()
   use M_fdata, only: nssh, rcutoff
   implicit none
   integer :: iatom,jatom,mbeta,num_neigh,num_neigh_vdw,in1,imu,in2,neighcount
-	real :: distance2,rcutoff_j, rcutoff_i,distance,range2,rc_max
+	real(8) :: distance2,rcutoff_j, rcutoff_i,distance,range2,rc_max
   integer :: mbeta_max2
   integer :: neigh_max_old, ii, jj
   if (icluster .eq. 1) mbeta_max = 0
